+++ src/config.cpp
@@ -120,6 +120,8 @@ int parse_arguments(int argc, char* argv[], BenchmarkConfig& config) {
           throw std::invalid_argument(Messages::error_missing_value("-output"));
       } else if (arg == "-patterns") {
         config.run_patterns = true;
+      } else if (arg == "-non-cacheable") {
+        config.use_non_cacheable = true;
       } else if (arg == "-h" || arg == "--help") {
         print_usage(argv[0]);
         return EXIT_SUCCESS;  // Special return value for help